# openapi.yaml
openapi: 3.0.3
info:
  title: Aetheria API
  description: |
    REST API for the Aetheria shared-world mobile MMORPG.
    All authenticated endpoints expect `Authorization: Bearer <accessToken>`.
    All state-mutating endpoints are server-authoritative.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://api.aetheria.example
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
            examples:
              hero:
                value:
                  email: "hero@example.com"
                  password: "correcthorsebattery"
                  displayName: "Lyra"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '409': { description: Email or display name already taken }
  /auth/login:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResult' }
        '401': { description: Invalid credentials }
  /auth/google:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [idToken]
              properties:
                idToken: { type: string, description: Google id_token (JWT) }
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthResult' } } }
  /auth/refresh:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties: { refreshToken: { type: string } }
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/AuthResult' } } }
...
  /characters:
    get:
      tags: [characters]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Character' }
    post:
      tags: [characters]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCharacterRequest' }
            examples:
              warrior:
                value: { name: "Kael", class: "warrior" }
      responses:
        '201':
          content: { application/json: { schema: { $ref: '#/components/schemas/Character' } } }
  /characters/{id}:
    get:
      tags: [characters]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/Character' } } }
  /characters/{id}/region:
    put:
      tags: [characters]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [regionId]
              properties: { regionId: { type: string } }
      responses: { '200': { description: Region updated } }
  /characters/{characterId}/inventory:
    get:
      tags: [inventory]
      parameters: [{ in: path, name: characterId, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/InventoryEntry' }
  /characters/{characterId}/inventory/equip:
    post:
      tags: [inventory]
      parameters: [{ in: path, name: characterId, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [inventoryItemId]
              properties: { inventoryItemId: { type: string, format: uuid } }
      responses: { '200': { description: Equipped } }
  /characters/{characterId}/inventory/unequip:
    post:
      tags: [inventory]
      parameters: [{ in: path, name: characterId, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [slot]
              properties: { slot: { type: string } }
      responses: { '200': { description: Unequipped } }
...
  /combat/start:
    post:
      tags: [combat]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartEncounterRequest' }
      responses:
        '201':
          content: { application/json: { schema: { $ref: '#/components/schemas/EncounterStarted' } } }
  /combat/{encounterId}/action:
    post:
      tags: [combat]
      parameters: [{ in: path, name: encounterId, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CombatActionRequest' }
            examples:
              attack: { value: { action: "attack" } }
              skill:  { value: { action: "skill", skillId: "power_strike" } }
              flee:   { value: { action: "flee" } }
      responses:
        '200':
          content: { application/json: { schema: { $ref: '#/components/schemas/CombatActionResult' } } }
  /combat/logs/{characterId}:
    get:
      tags: [combat]
      parameters: [{ in: path, name: characterId, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Recent fight log } }
  /guilds:
    post:
      tags: [guilds]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, tag]
              properties:
                name: { type: string, minLength: 3, maxLength: 24 }
                tag: { type: string, minLength: 2, maxLength: 6, pattern: '^[A-Z0-9]+$' }
                description: { type: string }
      responses: { '201': { description: Created } }
...
  /guilds/{id}:
    get:
      tags: [guilds]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Guild detail } }
  /guilds/{id}/join:
    post:
      tags: [guilds]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Joined } }
...
  /guilds/promote:
    post:
      tags: [guilds]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [playerId, rank]
              properties:
                playerId: { type: string, format: uuid }
                rank: { type: string, enum: [officer, member] }
      responses: { '200': { description: Updated } }
  /chat/send:
    post:
      tags: [chat]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [channel, body]
              properties:
                channel: { type: string, enum: [global, guild, dm] }
                recipientDisplayName: { type: string }
                body: { type: string, maxLength: 280 }
      responses: { '201': { description: Sent } }
  /chat/global:
    get:
      tags: [chat]
      parameters: [{ in: query, name: before, schema: { type: string, format: date-time } }]
      responses: { '200': { description: Recent messages } }
  /chat/guild/{guildId}:
    get:
      tags: [chat]
      parameters:
        - { in: path, name: guildId, required: true, schema: { type: string, format: uuid } }
        - { in: query, name: before, schema: { type: string, format: date-time } }
      responses: { '200': { description: Recent messages } }
  /chat/dm/{otherDisplayName}:
    get:
      tags: [chat]
      parameters:
        - { in: path, name: otherDisplayName, required: true, schema: { type: string } }
        - { in: query, name: before, schema: { type: string, format: date-time } }
      responses: { '200': { description: DM thread } }
...
  /leaderboards/{board}:
    get:
      tags: [leaderboards]
      security: []
      parameters:
        - { in: path, name: board, required: true, schema: { type: string, enum: [xp_total, gold_total] } }
        - { in: query, name: limit, schema: { type: integer, default: 50, maximum: 100 } }
      responses: { '200': { description: Top characters } }
  /leaderboards/{board}/rank/{characterId}:
    get:
      tags: [leaderboards]
      parameters:
        - { in: path, name: board, required: true, schema: { type: string, enum: [xp_total, gold_total] } }
        - { in: path, name: characterId, required: true, schema: { type: string, format: uuid } }
      responses: { '200': { description: This character's rank } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 10 }
        displayName: { type: string, minLength: 3, maxLength: 24 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AuthResult:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user:
          type: object
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
            displayName: { type: string }
...
    Character:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        class: { type: string, enum: [warrior, ranger, mage] }
        level: { type: integer }
        xp: { type: integer }
        xpToNext: { type: integer }
        skillPoints: { type: integer }
        stats: { $ref: '#/components/schemas/Stats' }
        equipment: { type: object, additionalProperties: { type: string } }
        currentRegionId: { type: string, nullable: true }
    Stats:
      type: object
      properties:
        health: { type: integer }
        attack: { type: integer }
        defense: { type: integer }
        agility: { type: integer }
        magic: { type: integer }
    CreateCharacterRequest:
      type: object
      required: [name, class]
      properties:
        name: { type: string, minLength: 2, maxLength: 24 }
        class: { type: string, enum: [warrior, ranger, mage] }
...
    InventoryEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        itemId: { type: string }
        name: { type: string }
        kind: { type: string }
        rarity: { type: string }
        slot: { type: string, nullable: true }
        stats: { type: object, additionalProperties: { type: number } }
        quantity: { type: integer }
        isEquipped: { type: boolean }
...
    StartEncounterRequest:
      type: object
      required: [characterId, regionId]
      properties:
        characterId: { type: string, format: uuid }
        regionId: { type: string }
    EncounterStarted:
      type: object
      properties:
        encounterId: { type: string, format: uuid }
        enemy:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            level: { type: integer }
            archetype: { type: string }
            stats: { $ref: '#/components/schemas/Stats' }
        player: { $ref: '#/components/schemas/Stats' }
        regionId: { type: string }
    CombatActionRequest:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [attack, skill, defend, use_item, flee] }
        skillId: { type: string, enum: [power_strike, arcane_bolt, aimed_shot] }
        itemId: { type: string }
    CombatActionResult:
      type: object
      properties:
        status: { type: string, enum: [ongoing, victory, defeat, fled] }
        encounterId: { type: string, format: uuid }
        step:
          type: object
          properties:
            round: { type: integer }
            playerAction: { type: string }
            playerDamage: { type: integer }
            enemyAction: { type: string }
            enemyDamage: { type: integer }
            playerHp: { type: integer }
            enemyHp: { type: integer }
        state:
          type: object
          properties:
            player: { $ref: '#/components/schemas/Stats' }
            enemy: { $ref: '#/components/schemas/Stats' }
        rewards:
          type: object
          properties:
            xp: { type: integer }
            gold: { type: integer }
            loot:
              type: array
              items: { type: object, properties: { item_id: { type: string }, quantity: { type: integer } } }
...